rubocop -a

+++ app/controllers/api/games_controller.rb
@@ -1,9 +1,10 @@
-class Api::GamesController < Api::ApiController
+# frozen_string_literal: true
 
+class Api::GamesController < Api::ApiController
   skip_before_action :verify_authenticity_token, only: %i[destroy cancel attack skip]
-  
+
   respond_to :json
-  
+
   def index
     render json: current_api_user.active_games.ordered
   end
@@ -12,8 +13,7 @@ class Api::GamesController < Api::ApiController
     render json: { count: current_api_user.active_games.count }
   end
 
-  def next
-  end
+  def next; end
 
   def skip
     status = -1
@@ -24,24 +24,22 @@ class Api::GamesController < Api::ApiController
     end
     render json: { status: status }
   end
-  
+
   def destroy
     status = -1
     game = Game.find_game(current_api_user, params[:id])
-    if game && game.winner
+    if game&.winner
       status = game.id
       if game.user_1 == current_api_user
         if game.user_2.bot
-          game.destroy()
+          game.destroy
         else
           game.update_attributes(del_user_1: true)
         end
       elsif game.user_2 == current_api_user
         game.update_attributes(del_user_2: true)
       end
-      if game.del_user_1 && game.del_user_2
-        game.destroy
-      end
+      game.destroy if game.del_user_1 && game.del_user_2
     end
     render json: { status: status }
   end
@@ -90,7 +88,7 @@ class Api::GamesController < Api::ApiController
     status = game && game.turn == current_api_user ? 1 : -1
     render json: { status: status }
   end
-  
+
   def show
     game = Game.find_game(current_api_user, params[:id])
     if game
@@ -98,15 +96,15 @@ class Api::GamesController < Api::ApiController
       moves_user = current_api_user == game.user_1 ? game.user_2 : game.user_1
       layouts_user = current_api_user == game.user_1 ? game.user_1 : game.user_2
       render json: {
-               game: klass.new(game, {}).as_json,
-               layouts: klass.new(game.layouts.where(user: layouts_user).ordered, {}).as_json,
-               moves: klass.new(game.moves_for_user(moves_user).ordered, {}).as_json
-             }
+        game: klass.new(game, {}).as_json,
+        layouts: klass.new(game.layouts.where(user: layouts_user).ordered, {}).as_json,
+        moves: klass.new(game.moves_for_user(moves_user).ordered, {}).as_json
+      }
     else
       render json: { error: 'game not found' }, status: :not_found
     end
   end
-  
+
   def opponent
     game = Game.find_game(current_api_user, params[:id])
     if game
@@ -114,10 +112,10 @@ class Api::GamesController < Api::ApiController
       moves_user = current_api_user == game.user_1 ? game.user_1 : game.user_2
       layouts_user = current_api_user == game.user_1 ? game.user_2 : game.user_1
       render json: {
-               game: klass.new(game, {}).as_json,
-               layouts: klass.new(game.layouts.where(user: layouts_user, sunk: true).ordered, {}).as_json,
-               moves: klass.new(game.moves_for_user(moves_user).ordered, {}).as_json
-             }
+        game: klass.new(game, {}).as_json,
+        layouts: klass.new(game.layouts.where(user: layouts_user, sunk: true).ordered, {}).as_json,
+        moves: klass.new(game.moves_for_user(moves_user).ordered, {}).as_json
+      }
     else
       render json: { error: 'game not found' }, status: :not_found
     end
@@ -133,12 +131,11 @@ class Api::GamesController < Api::ApiController
         shots = JSON.parse(params[:s]).slice(0, game.five_shot ? 5 : 1)
         shots.each do |s|
           move = game.moves.for_user(current_api_user).where(x: s['x'], y: s['y']).first
-          if move.nil?
-            layout_user = game.user_1 == current_api_user ? game.user_2 : game.user_1
-            layout = game.is_hit?(layout_user, s['x'], s['y'])
-            Move.create!(game: game, user: current_api_user, x: s['x'], y: s['y'], layout: layout)
-            layout.check_sunk if layout
-          end
+          next unless move.nil?
+          layout_user = game.user_1 == current_api_user ? game.user_2 : game.user_1
+          layout = game.is_hit?(layout_user, s['x'], s['y'])
+          Move.create!(game: game, user: current_api_user, x: s['x'], y: s['y'], layout: layout)
+          layout&.check_sunk
         end
         status = 1
         game.next_turn
@@ -155,7 +152,7 @@ class Api::GamesController < Api::ApiController
               end
               (5 - opponent.strength).times do
                 count += 1
-                log("  count: #{count}")                
+                log("  count: #{count}")
                 game.attack_random_ship(opponent, current_api_user)
               end
             else
@@ -165,7 +162,7 @@ class Api::GamesController < Api::ApiController
             game.next_turn if game.winner.nil?
             opponent.update_attributes(activity: opponent.activity + 1)
           else
-            current_api_user.update_attributes(activity: current_api_user.activity + 1)            
+            current_api_user.update_attributes(activity: current_api_user.activity + 1)
           end
         end
       end
@@ -174,5 +171,4 @@ class Api::GamesController < Api::ApiController
       render json: { error: 'game not found' }, status: :not_found
     end
   end
-  
 end
